Fix for returning to app from background in example.
When pulling down the status bar on iOS, the app will resign, however when returning to the app by swiping up the status bar, the notification of returning is not sent on the Notification Center but rather on the app delegate.
This can be monitored using React Native's AppState and handled appropriately in apps as demonstrated in the iOS example.